Convert to monorepo layout.
The thinking here is that these microservices will need code (scripts
and docker-compose files and the like) that will tie them together on
top of the code of the services themselves. A monorepo seems to be the
way to go for that.

The project formerly known as "server" is now located in the api
directory. The project formerly know as "frontend" is now in the web
folder.